Ensure proper alignment of elements in 'library' mode (fixes #27)
Was a complete crapshoot if you got a compile that wouldn't crash. While the issue only
mentions debug mode, this also happens in release mode sometimes. This was due to where
templates were being instantiated. Due to the restore functionality explicitly instantiating
all possible combinations of baobzi objects, the statics were being initialized given the
compile flags of the main baobzi object, which is compiled for generic architecture. Eigen
picks up on this and instantiates the objects with its generic alignment, which is only 16
bytes. AVX512 expects 64 bytes, unless you're using unaligned loads. Eigen was sometimes using
aligned loads on these unaligned objects for specific 'Orders' or just by chance, causing a
segfault. Since these objects were aligned on the stack, it would crash consistently in the
exact same way.

Other crashes were also happening on the heap (notably in destructors) due to the same
problem. Now all data AND instructions are generated with the same flags, which should
hopefully fix this issue.