Add minimum_leaf_fraction parameter
Threshold to decide if baobzi should artificially turn some leaves back into parents and start
the subtrees at a lower depth.

If the percentage of nodes that are leaves at a given depth are less than this parameter, then
make all nodes at this depth a parent and go one deeper. This costs memory, but can speed up
lookup by making node traversal closer to constant in time. This also greatly increases the
cache miss rate in some cases, and so should be tested carefully.

Leaving this parameter at its default of zero will keep the old behavior of setting the root of
the subtree at the first depth that has a single leaf.